wip: local context / recursive handling
