+++ distage/distage-core-api/src/main/scala/izumi/distage/model/exceptions/planning/InjectorFailed.scala
@@ -4,3 +4,4 @@ import izumi.distage.model.definition.errors.DIError
 import izumi.distage.model.exceptions.DIException
 
 class InjectorFailed(message: String, val errors: List[DIError]) extends DIException(message)
+
