+++ distage/distage-core-api/src/main/scala/izumi/distage/model/exceptions/planning/LocalContextVerificationFailed.scala
@@ -1,0 +1,7 @@
+package izumi.distage.model.exceptions.planning
+
+import izumi.distage.model.definition.ImplDef
+import izumi.distage.model.exceptions.DIException
+import izumi.distage.model.planning.PlanIssue
+
+class LocalContextVerificationFailed(message: String, val context: ImplDef.ContextImpl, val issues: Set[PlanIssue]) extends DIException(message)
