+++ distage/distage-core/src/main/scala/izumi/distage/model/exceptions/PlanVerificationException.scala
@@ -1,9 +1,9 @@
 package izumi.distage.model.exceptions
 
-import izumi.distage.planning.solver.PlanVerifier
+import izumi.distage.model.planning.PlanIssue
 import izumi.distage.planning.solver.PlanVerifier.PlanVerifierResult
 import izumi.fundamentals.collections.nonempty.NonEmptySet
 
 class PlanVerificationException(message: String, val cause: Either[Throwable, PlanVerifierResult.Incorrect]) extends DIException(message, cause.left.toOption.orNull) {
-  def issues: Option[NonEmptySet[PlanVerifier.PlanIssue]] = cause.toOption.flatMap(_.issues)
+  def issues: Option[NonEmptySet[PlanIssue]] = cause.toOption.flatMap(_.issues)
 }
