+++ distage/distage-framework/.jvm/src/main/scala/izumi/distage/framework/model/PlanCheckResult.scala
@@ -1,8 +1,8 @@
 package izumi.distage.framework.model
 
 import izumi.distage.framework.model.exceptions.PlanCheckException
+import izumi.distage.model.planning.PlanIssue
 import izumi.distage.model.reflection.DIKey
-import izumi.distage.planning.solver.PlanVerifier
 import izumi.distage.planning.solver.PlanVerifier.PlanVerifierResult
 import izumi.distage.plugins.load.LoadedPlugins
 import izumi.fundamentals.collections.nonempty.NonEmptySet
@@ -14,7 +14,7 @@ sealed abstract class PlanCheckResult {
   def maybeError: Option[Either[Throwable, PlanVerifierResult.Incorrect]]
   def maybeErrorMessage: Option[String]
 
-  final def issues: Option[NonEmptySet[PlanVerifier.PlanIssue]] = maybeError.flatMap(_.toOption.flatMap(_.issues))
+  final def issues: Option[NonEmptySet[PlanIssue]] = maybeError.flatMap(_.toOption.flatMap(_.issues))
   final def verificationPassed: Boolean = maybeError.isEmpty
   final def verificationFailed: Boolean = maybeError.isDefined
